NVIDIA_Cloud/requirements.txt
openai 
python-dotenv 
langchain_nvidia_ai_endpoints 
langchain_community 
faiss-cpu 
streamlit 
pypdf 